Play with transducer
I just tried implement transducers in Racket.

1 - We can skip the "reduced" concept because Racket has continuations,
so it's trivial to interrupt the call sequence.
2 - "Reduce" in Racket (foldl) has oposite order of arguments as
Clojure, also "cons" and others, so it makes sense to invert that in
transducers too
3 - Missing generic collections is Racket always get in the way. If I
reimplement map to work as Clojure, it gets restricted to lists. If I
try to use "sequence-map", I can't materialize the sequence.