Day 11 - first star
What I like:
- Operating with matrices separate the logic from the logic for
iterating over it.
- math/array lib has some clever functions.
- Learn typed Racket was fun.

What I don't like:
- Untyped math/array is REALLY slow.
- Typed Racket error are still a bit enigmatic.
- Typed math/array doesn't play nicely with `check-equal?`
- I can't shake the feeling that plain iteration would run faster than
using that library. That's still something to be tested.
- I had to implement `convolve`. Which, of course, is a half-baked
version of it.

Notes: I started with an untype implementation. It worked perfectly but
was unbelievably slow. So I decided to change everything to typed Racket.

It was a good decision, I've learned a little bit of it and the program
runs a lot faster now.

A little rant: static types always feel comfortable to code. I can see
myself thinking "could I make this type even more precise?", and Typed
Racket has lots of tools for that. On the other hand, it consumes a lot
of time for not so much benefit. For "safety", it's a poor man's version
of unit tests or contracts, missing most of the common programming
errors ("off by one", "out of boundaries", "empty/optional values", and
any logic error). For "speed", it shifts the burden from the compiler to
the programmer, which seems a questionable benefit unless you _really_
need that speed.

On top of that, when we move to collections, types become too complex
for easy reasoning. Covariance and contravariance on top of polymorphic
types made simple generic code the hell on earth.

Rant is over, sorry for that.