Day 11 - refactor before 2nd star
Here I'm generalizing `array-convolve` to be able to work with different
arrays and kernel sizes. Impressively, the code was slightly shorter,
although more difficult to understand.

Now I can implement a kernel with the same size of the matrix. But I'm
still not sure how to implement the "vision blocking" part. We'll see.